refactor: inline card list container into page
